[MIL_OPCOM] Implemented feature "custom objective" https://dev-heaven.net/issues/73599
Place a location-gamelogic of any type on the map and sync it to the OPCOM modules that should know about it.
optionally: put in init field of the logic -> this setvariable ["size",150] to set the size and this setvariable ["priority", 500] to set priority
optionally: [_OPCOM,"addObjective",[_id, _pos, _size]] call ALiVE_fnc_OPCOM; for mission makers ( f.e. _opcom = OPCOM_INSTANCES select 0);