[MIL OPCOM] Added ALiVE_fnc_OPCOMJoinNearestGroup and ALiVE_fnc_OPCOMJoinObjective!
ALiVE_fnc_JoinNearestGroup:
Usage: Execute that function locally on any players machine
[player,"attacking"] call ALiVE_fnc_OPCOMjoinNearestGroup (possible: "attacking"/"defending")

This will select the most important objective that is currently being attacked/defended by your forces and join you to an AI group that is on and assault/defend mission! This will set you directly in the action!

ALiVE_fnc_OPCOMJoinObjective:
Usage: Execute that function locally on any players machine
[player,"attacking"] call ALiVE_fnc_OPCOMjoinObjective (possible: "attacking"/"defending")

This will display all objectives that are currently beeing attacked/defended, lets you select the objective of choice and HALO drops your group to an AI group that is currently ordered to assault/defend this objective!